Function simple_arith_circuit::builder::add_equality_check

pub fn add_equality_check<F: FiniteField>(
    circuit: Circuit<F>,
    value: F
) -> Circuit<F>

Adds an equality check to circuit such that the new circuit outputs zero if the output of the old circuit is value.

Errors

Errors out if the circuit does not have exactly one output wire.